requirements.txt
Flask
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
python-docx
docx
google-auth
